#27 add donate link
added dontate link to app
updated readme with dontation info and banner
